backend/requirements.txt
Django>=6.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
psycopg2-binary>=2.9
python-decouple>=3.8
boto3>=1.34
Pillow>=10.0
